Move Batik utils to extensions
